+++ packages/cli/lib/services/dryrun.service.ts
@@ -1,17 +1,20 @@
 import promptly from 'promptly';
 import chalk from 'chalk';
 
-import type { NangoConnection } from '@nangohq/shared';
+import type { NangoConnection, NangoProps, ScriptExecutorInterface, RunScriptOptions, RunnerOutput } from '@nangohq/shared';
 import type { Metadata, ScriptFileType } from '@nangohq/types';
-import { SyncType, cloudHost, stagingHost, SyncRunService, NangoError, localFileService } from '@nangohq/shared';
+import { cloudHost, stagingHost, NangoError, localFileService, validateData, NangoSync, formatScriptError, ActionError } from '@nangohq/shared';
 import type { GlobalOptions } from '../types.js';
 import { parseSecretKey, printDebug, hostport, getConnection, getConfig } from '../utils.js';
 import { compileAllFiles } from './compile.service.js';
-import integrationService from './local-integration.service.js';
-import type { RecordsServiceInterface } from '@nangohq/shared/lib/services/sync/run.service.js';
 import { parse } from './config.service.js';
 import { loadSchemaJson } from './model.service.js';
 import { displayValidationError } from '../utils/errors.js';
+import * as vm from 'vm';
+import * as url from 'url';
+import * as crypto from 'crypto';
+import * as zod from 'zod';
+import { Buffer } from 'buffer';
 
 interface RunArgs extends GlobalOptions {
     sync: string;
@@ -24,7 +27,7 @@ interface RunArgs extends GlobalOptions {
     optionalProviderConfigKey?: string;
 }
 
-export class DryRunService {
+export class DryRunService implements ScriptExecutorInterface {
     fullPath: string;
     validation: boolean;
     environment?: string;
@@ -248,38 +251,14 @@ export class DryRunService {
             messages: []
         };
 
-        // dry-run mode does not read or write to the records database
-        // so we can safely mock the records service
-        const recordsService: RecordsServiceInterface = {
-            markNonCurrentGenerationRecordsAsDeleted: ({
-                connectionId: _connectionId,
-                model: _model,
-                syncId: _syncId,
-                generation: _generation
-            }: {
-                connectionId: number;
-                model: string;
-                syncId: string;
-                generation: number;
-                // eslint-disable-next-line @typescript-eslint/require-await
-            }): Promise<string[]> => {
-                return Promise.resolve([]);
-            }
-        };
-
         const jsonSchema = loadSchemaJson({ fullPath: this.fullPath });
         if (!jsonSchema) {
             console.log(chalk.red('Failed to load schema.json'));
             return;
         }
 
-        const syncRun = new SyncRunService({
-            integrationService,
-            recordsService,
-            dryRunService: new DryRunService({ environment, returnOutput: true, fullPath: this.fullPath, validation: this.validation }),
-            writeToDb: false,
-            nangoConnection,
-            syncConfig: {
+        try {
+            const syncConfig = {
                 sync_name: syncName,
                 file_location: '',
                 models: syncInfo?.output || [],
@@ -297,30 +276,45 @@ export class DryRunService {
                 models_json_schema: jsonSchema,
                 created_at: new Date(),
                 updated_at: new Date()
-            },
-            provider,
-            input: normalizedInput as object,
-            isAction: syncInfo?.type === 'action',
-            isPostConnectionScript,
-            syncId: 'abc',
-            syncJobId: -1,
-            syncType: SyncType.INITIAL,
-            loadLocation: './',
-            debug,
-            logMessages,
-            stubbedMetadata,
-            runnerFlags: {
-                validateActionInput: this.validation, // irrelevant for cli
-                validateActionOutput: this.validation, // irrelevant for cli
-                validateSyncRecords: this.validation,
-                validateSyncMetadata: false
-            }
-        });
+            };
+            const nangoProps: NangoProps = {
+                host: process.env['NANGO_HOSTPORT'],
+                connectionId: nangoConnection.connection_id,
+                environmentId: nangoConnection.environment_id,
+                environmentName: environment,
+                providerConfigKey: nangoConnection.provider_config_key,
+                provider,
+                secretKey: process.env['NANGO_SECRET_KEY'] || '',
+                nangoConnectionId: nangoConnection.id as number,
+                syncId: 'dryrun-sync',
+                lastSyncDate: lastSyncDate as Date,
+                dryRun: true,
+                logMessages,
+                stubbedMetadata,
+                syncConfig,
+                dryRunService: new DryRunService({ environment, returnOutput: true, fullPath: this.fullPath, validation: this.validation }),
+                runnerFlags: {
+                    validateActionInput: this.validation, // irrelevant for cli
+                    validateActionOutput: this.validation, // irrelevant for cli
+                    validateSyncRecords: this.validation,
+                    validateSyncMetadata: false
+                }
+            };
+            const isAction = syncInfo?.type === 'action';
+            const isWebhook = false;
+            const isInvokedImmediately = Boolean(isAction || isWebhook || isPostConnectionScript);
 
-        try {
             console.log('---');
-            const secretKey = process.env['NANGO_SECRET_KEY'];
-            const results = await syncRun.run(lastSyncDate, true, secretKey, process.env['NANGO_HOSTPORT']);
+            const results = await this.runScript({
+                syncName,
+                nangoProps,
+                isInvokedImmediately,
+                isWebhook,
+                syncId: nangoProps.syncId as string,
+                optionalLoadLocation: './',
+                input: normalizedInput,
+                writeToDb: false
+            });
             console.log('---');
 
             if (results.error) {
@@ -352,28 +346,28 @@ export class DryRunService {
                 }
             }
 
-            if (syncRun.logMessages && syncRun.logMessages.messages.length > 0) {
-                const logMessages = syncRun.logMessages.messages;
+            if (logMessages && logMessages.messages.length > 0) {
+                const messages = logMessages.messages;
                 let index = 0;
                 const batchCount = 10;
 
                 const displayBatch = () => {
-                    for (let i = 0; i < batchCount && index < logMessages.length; i++, index++) {
-                        const logs = logMessages[index];
+                    for (let i = 0; i < batchCount && index < messages.length; i++, index++) {
+                        const logs = messages[index];
                         console.log(chalk.yellow(JSON.stringify(logs, null, 2)));
                         resultOutput.push(JSON.stringify(logs, null, 2));
                     }
                 };
 
-                console.log(chalk.yellow(`The dry run would produce the following results: ${JSON.stringify(syncRun.logMessages.counts, null, 2)}`));
-                resultOutput.push(`The dry run would produce the following results: ${JSON.stringify(syncRun.logMessages.counts, null, 2)}`);
+                console.log(chalk.yellow(`The dry run would produce the following results: ${JSON.stringify(logMessages.counts, null, 2)}`));
+                resultOutput.push(`The dry run would produce the following results: ${JSON.stringify(logMessages.counts, null, 2)}`);
                 console.log(chalk.yellow('The following log messages were generated:'));
                 resultOutput.push('The following log messages were generated:');
 
                 displayBatch();
 
-                while (index < syncRun.logMessages.messages.length) {
-                    const remaining = syncRun.logMessages.messages.length - index;
+                while (index < logMessages.messages.length) {
+                    const remaining = logMessages.messages.length - index;
                     const confirmation = await promptly.confirm(
                         `There are ${remaining} logs messages remaining. Would you like to see the next 10 log messages? (y/n)`
                     );
@@ -392,6 +386,135 @@ export class DryRunService {
             process.exit(0);
         } catch {
             process.exit(1);
+        }
+    }
+
+    async runScript({ syncName, nangoProps, isInvokedImmediately, isWebhook, optionalLoadLocation, input }: RunScriptOptions): Promise<RunnerOutput> {
+        const nango = new NangoSync(nangoProps);
+        try {
+            await nango.log(`Executing -> integration:"${nangoProps.provider}" script:"${syncName}"`);
+
+            const script: string | null = localFileService.getIntegrationFile(syncName, nangoProps.providerConfigKey, optionalLoadLocation);
+            const isAction = isInvokedImmediately && !isWebhook;
+
+            if (!script) {
+                const content = `Unable to find script file for "${syncName}"`;
+                return { success: false, error: new NangoError(content, 500), response: null };
+            }
+
+            try {
+                const wrappedScript = `
+                    (function() {
+                        var module = { exports: {} };
+                        var exports = module.exports;
+                        ${script}
+                        return module.exports;
+                    })();
+                `;
+
+                const scriptObj = new vm.Script(wrappedScript);
+                const sandbox = {
+                    console,
+                    require: (moduleName: string) => {
+                        switch (moduleName) {
+                            case 'url':
+                                return url;
+                            case 'crypto':
+                                return crypto;
+                            case 'zod':
+                                return zod;
+                            default:
+                                throw new Error(`Module '${moduleName}' is not allowed`);
+                        }
+                    },
+                    Buffer,
+                    setTimeout
+                };
+
+                const context = vm.createContext(sandbox);
+                const scriptExports: any = scriptObj.runInContext(context);
+
+                if (!scriptExports.default || !(typeof scriptExports.default === 'function')) {
+                    const content = `There is no default export that is a function for ${syncName}`;
+                    return { success: false, error: new NangoError(content, 500), response: null };
+                }
+
+                if (isAction) {
+                    // Validate action input against json schema
+                    const valInput = validateData({
+                        version: nangoProps.syncConfig.version || '1',
+                        input: input,
+                        modelName: nangoProps.syncConfig.input,
+                        jsonSchema: nangoProps.syncConfig.models_json_schema
+                    });
+                    if (Array.isArray(valInput)) {
+                        await nango.log('Invalid action input. Use `--validation` option to see the details', { level: 'warn' });
+                        if (nangoProps.runnerFlags.validateActionInput) {
+                            return {
+                                success: false,
+                                response: null,
+                                error: new NangoError('invalid_action_input', { data: input, validation: valInput, model: nangoProps.syncConfig.input })
+                            };
+                        }
+                    }
+
+                    const output = await scriptExports.default(nango, input);
+
+                    // Validate action output against json schema
+                    const modelNameOutput = nangoProps.syncConfig.models.length > 0 ? nangoProps.syncConfig.models[0] : undefined;
+                    const valOutput = validateData({
+                        version: nangoProps.syncConfig.version || '1',
+                        input: output,
+                        modelName: modelNameOutput,
+                        jsonSchema: nangoProps.syncConfig.models_json_schema
+                    });
+                    if (Array.isArray(valOutput)) {
+                        await nango.log('Invalid action output. Use `--validation` option to see the details', { level: 'warn' });
+                        if (nangoProps.runnerFlags.validateActionOutput) {
+                            return {
+                                success: false,
+                                response: null,
+                                error: new NangoError('invalid_action_output', { data: output, validation: valOutput, model: modelNameOutput })
+                            };
+                        }
+                    }
+
+                    return { success: true, error: null, response: output };
+                }
+
+                const results = await scriptExports.default(nango);
+                return { success: true, error: null, response: results };
+            } catch (err) {
+                if (err instanceof ActionError) {
+                    return {
+                        success: false,
+                        error: {
+                            type: err.type,
+                            payload: err.payload || {},
+                            status: 500
+                        },
+                        response: null
+                    };
+                } else if (err instanceof NangoError) {
+                    return { success: false, error: err, response: null };
+                }
+
+                let errorType = 'sync_script_failure';
+                if (isWebhook) {
+                    errorType = 'webhook_script_failure';
+                } else if (isInvokedImmediately) {
+                    errorType = 'action_script_failure';
+                }
+
+                return formatScriptError(err, errorType, syncName);
+            }
+        } catch (err) {
+            const errorMessage = JSON.stringify(err, ['message', 'name', 'stack'], 2);
+            const content = `The script failed to load for ${syncName} with the following error: ${errorMessage}`;
+
+            return { success: false, error: new NangoError(content, 500), response: null };
+        } finally {
+            await nango.log(`Done`);
         }
     }
 }
